requirements.txt
Django==5.2.10
psycopg2-binary==2.9.11
python-dotenv==1.2.1
gunicorn==23.0.0
iot-logging-schemas @ git+https://github.com/IoT-Hub-Alpha/logging-lib.git@dev
iot-auth @ git+https://github.com/IoT-Hub-Alpha/jwt-auth-lib.git@task-41-jwt-auth-lib
